Set a confidence in the search results.
For results in the repo, we set some arbitrary values in the
service for simplicity. We can move this responsibility to the repo
easily later if needed.

Results to the `find` query are returned sorted by confidence.

Add a debugString parameter to communicate information about the
match to the clients, e.g. how the company was found/matched, for
debugging.